1、服务器支持WebSocket通信； 2、优化脚本。

+++ src/server/mod_admin_extra/build.sh
@@ -1,9 +1,13 @@
 #!/bin/sh
 
-rm -rf ebin
+EJABBERD_PATH=/usr/lib/x86_64-linux-gnu/ejabberd
+
+if [ -d ebin ]; then
+	rm -rf ebin
+fi
+
 mkdir ebin
 
-erl -pa /lib/ejabberd/ebin -pz ebin -make || exit 1
+erl -pa ${EJABBERD_PATH}/ebin -pz ebin -make || exit 1
 
-# gen_mod.beam仅是中间文件
-rm -f ebin/gen_mod.beam
+cp -f ebin/* ${EJABBERD_PATH}/ebin
